Lessons from Brian Lonsdorf (@drboolean)
This is a collection of Ruby scripts which follow the "Professor Frisby
Introduces Composable Javascript" series available for free on
egghead.io:

https://egghead.io/courses/professor-frisby-introduces-composable-functional-javascript

This is basically a ruby port of the concepts and ideas shared, with
some small changes to follow more idiomatic ruby ideas.

It's fun, good to know and learn.

If you have thoughts, ideas, comments, please get in touch.